StatisticalGraphics.Band — Type

Band(args...)

Represent a Band plot with given arguments.

Keyword arguments

Required

lower: The lower values for the band plot. User can pass a column or a Float value. default: 0

upper: The upper values for the band plot. User can pass a column or a Float value. default: 0

x: The column to be used as x coordinate. User must pass either this or the y argument. default: 0

y: The column to be used as y coordinate. User must pass either this or the x argument. default: 0

Band Options

breaks: It causes a break in the line when a missing value is encountered. default: false

interpolate: The interplate function to use for drawing lines, e.g. :linear, :basis, :natural, :step, ... default: :linear

Grouping

group: The name of column for grouping observation. Each group of observations will create seperate band plot.

Band appearance

color: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4"), or pass a gradient color using the gradient() function. default: "#4682b4"

opacity: The mark opacity from 0 (transparent) to 1 (opaque). default: 0.5

Axes options

x2axis: When set to true, the top x-axis will be used for the current plot. default: false

y2axis: When set to true, the right y-axis will be used for the current plot. default: false

Legend

legend: User can pass a symbol to this keyword argument to indicate that more customisation will be passed for the legened of corresponding mark. User needs to provide the extra customisation via the Legend global keyword.

Miscellaneous

clip: Indicates if the marks should be clipped.

StatisticalGraphics.Bar — Type

Bar(args...)

Represent a Bar chart with given arguments.

Keyword arguments

Required

x: User should pass a single column for plotting the bar chart. User must pass either this or the y argument. default: 0

y: User should pass a single column for plotting the bar chart. User must pass either this or the x argument. default: 0

Bar options

baseline: The start value for drawing bars. default: 0

baselineresponse: Name/index of a numeric column which will be used to change the baseline of bars within each group. The function passed to baselinestat will be used to aggregate the values if there are more than one observation for a specific bar.

baselinestat: The function that will be used to aggregate values in column passed as baselineresponse.

colormodel: It specifies the color scheme to use for the marks. default: :diverging

colorresponse: Name/index of a numeric column which will be used to change the fill color of each bar based on its values. The function passed to colorstat will be used to aggregate the values if there are more than one observation for a specific bar.

colorstat: The function that will be used to aggregate values in column passed as colorresponse.

missingmode: Indicate how to handle missing values in category or group. 0 = nothing, 1 = no missing in category, 2 = no missing in group, 3 = no missing in category or group, 4 = no missing in category and group. default: 0

normalize: If true the bars will be normalized in each group. By default the total bar heights will be 1 in each group, however, user can pass customised function via normalizer to change this behaviour. default: false

normalizer: This function will be used to normalize bar height within each group. default: StatisticalGraphics.bar_normalizer

orderresponse: Name of a numeric column which will be used to change the order of bars. The function passed to orderstat will be used to aggregate the values if there are more than one observation for a specific bar. Note that the axis' order will override this.

orderstat: The function that will be used to aggregate values in column passed as orderresponse.

response: A numeric column which its aggregated values based on the stat keyword argument will be used to determine the height of each bar.

stat: A function for aggregating the response keyword argument. When response is passed the default value of the keword change to IMD.sum, however, user can pass any function to this argument. The function must accept two arguments f(format), and x the input values and return the aggregated values.

Grouping

group: A grouped bar chart will be created by passing a single column to this argument.

groupdisplay: Indicate how to display bars in each group. It can be :stack, :cluster, :step, or :none. default: :stack

grouporder: How to order values in each group. It can be :data, :ascending, or :descending. User may also pass a vector of group levels to dictate the orders. default: :ascending

groupspace: The space between bars inside each group when groupdisplay is in (:cluster, :step). default: 0.05

Bar appearance

barcorner: Corner radius for bars. (cornerRadiusTopLeft, cornerRadiusTopRight, cornerRadiusBottomLeft, cornerRadiusBottomRight). default: 0

barwidth: The bar width proportion with respect to the available space. It can be any number between 0 and 1. User can pass :nest too, which in this case the bar width will be automatically calculated for each group in such a way that the bar widths in each group will be smaller than the previous group. Users can pass nestfactor to control how fast they would like the bar width change for each group. default: 1

color: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4"), or pass a gradient color using the gradient() function. default: "#4682b4"

nestfactor: When the barwidth keyword is set to :nest this will control how much change should be applied to the current group barwidth compared to the previous one. By default this will be controlled automatically.

opacity: The mark opacity from 0 (transparent) to 1 (opaque). default: 1

outlinecolor: The mark's outline color. default: :white

outlinethickness: The mark outline thickness. default: 1

space: The space between bars. It can be any number between 0 and 1. default: 0.1

Bar labels

label: What information should be used for bar labels. It can be :none, :height, or :category. default: :none

labelalign: The text alignment.

labelalternate: Use automatic alogirhtms to adjust the labels for bar with negative heights. default: true

labelangle: The text angle.

labelbaseline: The text baseline.

labelcolor: The text color. User can also pass :group or :colorresponse to use the corresponding scale for coloring the text. default: :black

labeld3format: d3 format for labels. default: ""

labeldir: The direction of the text, i.e. :ltr or :rtl. default: :ltr

labelfont: The font name for displaying text.

labelfontweight: The font weight for displaying text, use 100 for thin font and 900 for the bold one.

labelitalic: If true the italic style will be used for displaying text.

labellimit: The maximum length of the text mark in pixels. The text value will be automatically truncated if the rendered size exceeds the limit.

labelloc: relative location of the label within each bar. It can be any number between 0 and 1, where 0.5 means middle of the bar. default: 0.5

labeloffset: The amount in pixel to offset the labels. default: 0

labelopacity: The text opacity. default: 1

labelpos: The position of labels within each bar. It can be :end, :start, or :middle. default: :end

labelsize: The font size for displaying text.

tooltip: A tooltip will be shown upon mouse hover. default: false

Axes options

x2axis: When set to true, the top x-axis will be used for the current plot. default: false

y2axis: When set to true, the right y-axis will be used for the current plot. default: false

Legend

legend: User can pass a symbol to this keyword argument to indicate that more customisation will be passed for the legened of corresponding mark. User needs to provide the extra customisation via the Legend global keyword.

Miscellaneous

clip: Indicates if the marks should be clipped.

StatisticalGraphics.BoxPlot — Type

BoxPlot(args...)

Represent a Box Plot with given arguments.

Keyword arguments

Required

x: User should pass multiple columns for plotting the comparative box plot, i.e. side by side box plot for passed columns. User must pass either this or the y argument. default: 0

y: User should pass multiple columns for plotting the comparative box plot, i.e. side by side box plot for passed columns. User must pass either this or the x argument. default: 0

BoxPlot options

categoryorder: How the category should be ordered, i.e. :ascending, :descending, :data. default: :ascending

missingmode: Indicate how to handle missing values in category or group. 0 = nothing, 1 = no missing in category. default: 0

outliers: If true the oultliers will be shown. The outliers are computed based on how many outliersfactor they are far from quartiles. default: false

outliersfactor: The factor to be used for computing outliers. default: 1.5

Grouping

category: A category column which indicates the box plot must be drawn within each category.

BoxPlot appearance

boxcorner: Corner radius for boxes. (cornerRadiusTopLeft, cornerRadiusTopRight, cornerRadiusBottomLeft, cornerRadiusBottomRight). default: 0

boxwidth: The box width, it must be a number between 0 and 1. default: 1

fencecolor: Specifiy the color for the fence lines. default: :black

fencewidth: Specifiy the total width for the fence lines. It must be a number between 0 and 1. default: 0.5

groupspace: The space between boxes inside each category. It must be a number between 0 and 1. default: 0.05

meansymbol: Specifiy the symbol to be used for the mean indicator. default: :diamond

meansymbolsize: Specifiy the symbol size to be used for the mean indicator. default: 40

mediancolor: Specifiy the color for the median indicator. default: :white

medianthickness: Specifiy the thickness for the median indicator. default: 1

medianwidth: The total width to be used for the median indicator. It must be a number between 0 and 1. default: 1

opacity: The mark opacity from 0 (transparent) to 1 (opaque). default: 1

outlinecolor: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4"). default: :white

outlinethickness: The mark outline thickness. default: 1

space: The space between boxes. It must be a number between 0 and 1. default: 0.1

whiskercolor: Specifiy the color for the whisker lines. default: :black

whiskerdash: Specifiy the dash style for the whisker lines. default: [3, 3]

whiskerthickness: Specifiy the thickness for the whisker lines. default: 1

Outliers appearance

outliercolor: Specifiy the color for the outliers symobls.

outlierjitter: Specifiy the jitter strength for the outliers symobls. default: 0

outlieropacity: Specifiy the mark opacity for the outliers symobls. default: 1

outlieroutlinecolor: Specifiy the outline color for the outliers symobls.

outliersymbol: Specifiy the symbol shape for the outliers symobls. default: :circle

outliersymbolsize: Specifiy the symbol size for the outliers symobls. default: 30

outlierthickness: Specifiy the outline thickness for the outliers symobls. default: 1

BoxPlot tooltip

tooltip: A tooltip will be shown upon mouse hover. default: false

Axes options

x2axis: When set to true, the top x-axis will be used for the current plot. default: false

y2axis: When set to true, the right y-axis will be used for the current plot. default: false

Legend

legend: User can pass a symbol to this keyword argument to indicate that more customisation will be passed for the legened of corresponding mark. User needs to provide the extra customisation via the Legend global keyword.

Miscellaneous

clip: Indicates if the marks should be clipped.

StatisticalGraphics.Bubble — Type

Bubble(args...)

Represent a Bubble plot with given arguments.

Keyword arguments

Required

x: The column to be used as x coordinate. default: 0

y: The column to be used as y coordinate. default: 0

Bubble Options

colorresponse: The column which its values will be used to determine the fill color of the marks.

maxsize: The maximum size of the bubbles.

minsize: The minimum size of the bubbles.

opacityresponse: The column which its values will be used to determine the opacity of the marks.

Grouping

group: The name of column for grouping observation. Each group of observations will create seperate bubble plot.

Bubble appearance

color: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4").

colormodel: It specifies the color scheme to use for the marks. default: :diverging

opacity: The mark opacity from 0 (transparent) to 1 (opaque). default: 1

outlinecolor: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4").

size: The symbol size. default: 50

thickness: The mark outline thickness. default: 1

Bubble Label

labelalgorithm: The algorithm for placing labels. default: :naive

labelanchor: The anchors for choosing the location of labels. default: [:top, :bottom, :left, :right]

labelangle: The text angle. default: 0

labelcolor: The label text color, it can also be :group or :colorresponse for choosing the color based on the points groups. default: :black

labeldir: The direction of the text, i.e. :ltr or :rtl. default: :ltr

labelfont: The font name for displaying text.

labelfontweight: The font weight for displaying text, use 100 for thin font and 900 for the bold one.

labelitalic: If true the italic style will be used for displaying text.

labellimit: The maximum length of the text mark in pixels. The text value will be automatically truncated if the rendered size exceeds the limit.

labelresponse: The column which its values will be used to label points.

labelsize: The font size for displaying text.

tooltip: A tooltip will be shown upon mouse hover. default: false

Axes options

x2axis: When set to true, the top x-axis will be used for the current plot. default: false

xshift: Shift the mark in direction of x. Useful for discrete type axes. default: 0

y2axis: When set to true, the right y-axis will be used for the current plot. default: false

yshift: Shift the mark in direction of y. Useful for discrete type axes. default: 0

Legend

legend: User can pass a symbol to this keyword argument to indicate that more customisation will be passed for the legened of corresponding mark. User needs to provide the extra customisation via the Legend global keyword.

Miscellaneous

clip: Indicates if the marks should be clipped.

StatisticalGraphics.Density — Type

Density(args...)

Represent a Density plot with given arguments.

Keyword arguments

Required

x: The column to be used as x coordinate. User must pass either this or the y argument. default: 0

y: The column to be used as y coordinate. User must pass either this or the x argument. default: 0

Density Options

baseline: The baseline for filling the curve. default: 0.0

bw: Band width to be used in the kernel density estimation.

interpolate: The interplate function to use for drawing lines, e.g. :linear, :basis, :natural, :step, ... default: :linear

npoints: The number of points for the grid calculation. default: 100

scale: user can pass any function to this option, the function must be in the form of fun(density; midpoints, npoints, samplesize, binwidth) , for :pdf the function is defined as f(x; args...) = x, for :count we compute the expected counts, f(x; args...) = x .* binwidth .* npoints , and for :cdf the function is (x; binwidth, args...) -> cumsum(x .* binwidth). default: :pdf

type: Type of density fit, i.e. :normal or :kernel. default: :normal

weights: The weighting function to be used when :kernel type is selected. default: :gaussian

Grouping

group: The name of column for grouping observation. Each group of observations will create seperate density plot.

Density appearance

color: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4"), or pass a gradient color using the gradient() function.

fillcolor: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4"), or pass a gradient color using the gradient() function.

filled: Indicate if the curve should be filled. default: true

fillopacity: The opacity value for the fill color. default: 0.5

opacity: The opacity value for the outline. default: 1

thickness: The mark outline thickness. default: 1

Axes options

x2axis: When set to true, the top x-axis will be used for the current plot. default: false

y2axis: When set to true, the right y-axis will be used for the current plot. default: false

Legend

legend: User can pass a symbol to this keyword argument to indicate that more customisation will be passed for the legened of corresponding mark. User needs to provide the extra customisation via the Legend global keyword.

Miscellaneous

clip: Indicates if the marks should be clipped.

StatisticalGraphics.Heatmap — Type

Heatmap(args...)

Represent a Heatmap (2D Histogram) with given arguments.

Keyword arguments

Required

x: The column to be used as x coordinate. User must pass this and the y argument. default: 0

y: The column to be used as y coordinate. User must pass this and the x argument. default: 0

Heatmap Options

bincountmethod: The number of points for the grid calculation in both direction. default: StatisticalGraphics.default_bin_heatmap

tooltip: A tooltip will be shown upon mouse hover. default: false

xbincount: The number of points for the grid calculation in x direction.

ybincount: The number of points for the grid calculation in y direction.

Heatmap appearance

colormodel: Define the color model to be used for the heatmap plot. default: :scheme => :viridis

opacity: The opacity value for the outline. default: 1

Axes options

x2axis: When set to true, the top x-axis will be used for the current plot. default: false

y2axis: When set to true, the right y-axis will be used for the current plot. default: false

Legend

legend: User can pass a symbol to this keyword argument to indicate that more customisation will be passed for the legened of corresponding mark. User needs to provide the extra customisation via the Legend global keyword.

Miscellaneous

clip: Indicates if the marks should be clipped.

StatisticalGraphics.Histogram — Type

Histogram(args...)

Represent a Histogram with given arguments.

Keyword arguments

Required

x: The column to be used as x coordinate. User must pass either this or the y argument. default: 0

y: The column to be used as y coordinate. User must pass either this or the x argument. default: 0

Histogram Options

midpoints: The location of midpoints. It can be a number to indicate the number of midpoints or a vector of midpoints. default: :Sturges

scale: The scale to use for the bar heights, e.g. :pdf, :cdf, :count. default: :pdf

Grouping

group: The name of column for grouping observation. Each group of observations will create seperate histogram plot.

Histogram appearance

color: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4"), or pass a gradient color using the gradient() function. default: "#4682b4"

opacity: The opacity value for the outline. default: 1

outlinecolor: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4"), or pass a gradient color using the gradient() function. default: :white

outlinethickness: The outline thickness. default: 1

space: Space between bars in pixel. default: 1

Axes options

x2axis: When set to true, the top x-axis will be used for the current plot. default: false

y2axis: When set to true, the right y-axis will be used for the current plot. default: false

Legend

legend: User can pass a symbol to this keyword argument to indicate that more customisation will be passed for the legened of corresponding mark. User needs to provide the extra customisation via the Legend global keyword.

Miscellaneous

clip: Indicates if the marks should be clipped.

StatisticalGraphics.Line — Type

Line(args...)

Represent a Line plot with given arguments.

Keyword arguments

Required

x: The column to be used as x coordinate. default: 0

y: The column to be used as y coordinate. default: 0

Line Options

breaks: It causes a break in the line when a missing value is encountered. default: false

interpolate: The interplate function to use for drawing lines, e.g. :linear, :basis, :natural, :step, ... default: :linear

Grouping

group: The name of column for grouping observation. Each group of observations will create seperate Line.

Line appearance

color: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4"), or pass a gradient color using the gradient() function. default: "#4682b4"

dash: The Line dash style. default: [0]

opacity: The mark opacity from 0 (transparent) to 1 (opaque). default: 1

thickness: The Line thickness. default: 1

Axes options

x2axis: When set to true, the top x-axis will be used for the current plot. default: false

xshift: Shift the mark in direction of x. Useful for discrete type axes. default: 0

y2axis: When set to true, the right y-axis will be used for the current plot. default: false

yshift: Shift the mark in direction of y. Useful for discrete type axes. default: 0

Legend

legend: User can pass a symbol to this keyword argument to indicate that more customisation will be passed for the legened of corresponding mark. User needs to provide the extra customisation via the Legend global keyword.

Miscellaneous

clip: Indicates if the marks should be clipped.

StatisticalGraphics.Pie — Type

Pie(args...)

Represent a Pie chart with given arguments.

Keyword arguments

Required

category: A category column which indicates each slice. default: 0

Pie options

endangle: The end angle in degree. default: 360

missingmode: Indicate how to handle missing values in category or group. 0 = nothing, 1 = no missing in category, 2 = no missing in group, 3 = no missing in category or group, 4 = no missing in category and group. default: 0

response: A numeric column which its aggregated values based on the stat keyword argument will be used to determine the radius of each slice.

sort: If true the slices will be sorted based on their total angle. default: false

startangle: The start angle in degree. default: 0

stat: A function for aggregating the response keyword argument. When response is passed the default value of the keword change to IMD.sum, however, user can pass any function to this argument. The function must accept two arguments f(format), and x the input values and return the aggregated values.

Grouping

group: A nested pie chart will be created by passing a single column to this argument.

groupspace: The space between nested pie. default: 0.01

Pie appearance

colormodel: It specifies the color scheme to use for the marks. default: :category

innerradius: The donut radius. It must be a number between 0 and 1. default: 0

opacity: The mark opacity from 0 (transparent) to 1 (opaque). default: 1

outerradius: The outer radius. It must be a number between 0 and 1. default: 1

outlinecolor: The mark's outline color. default: :white

outlinethickness: The mark outline thickness. default: 1

piecorner: The corner of slice in pixel. default: 0

space: The space between slices. It can be any number between 0 and 1. default: 0

Pie labels

decimal: Numbe of digits after the decimal when percentage is used. default: 1

label: What information should be used for slice labels. It can be :category, :percent, or :both.

labelalign: The text alignment. default: :center

labelangle: The text angle.

labelbaseline: The text baseline. default: :middle

labelcolor: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4"). default: :black

labeldir: The direction of the text, i.e. :ltr or :rtl. default: :ltr

labelfont: The font name for displaying text.

labelfontweight: The font weight for displaying text, use 100 for thin font and 900 for the bold one.

labelitalic: If true the italic style will be used for displaying text.

labellimit: The maximum length of the text mark in pixels. The text value will be automatically truncated if the rendered size exceeds the limit.

labelopacity: The text opacity. default: 1

labelpos: Vertical position of lables. default: 0.5

labelrotate: It rotates text 90 degree. default: false

labelsize: The font size for displaying text.

labelthreshold: Labels will be drop for the slices with smaller angle than :labelthreshold. default: 0.0

tooltip: A tooltip will be shown upon mouse hover. default: false

Legend

legend: User can pass a symbol to this keyword argument to indicate that more customisation will be passed for the legened of corresponding mark. User needs to provide the extra customisation via the Legend global keyword.

Miscellaneous

clip: Indicates if the marks should be clipped.

StatisticalGraphics.Polygon — Type

Polygon(args...)

Represent a Polygon with given arguments.

Keyword arguments

Required

id: The function draw a seperate polygon for each unique value of id.

x: The x coordinate of the polygon.

y: The y coordinate of the polygon.

Polygon Options

colorresponse: The column which its values will be used to determine the fill color of polygons.

interpolate: The interplate function to use for drawing lines, e.g. :linear, :basis, :natural, :step, ... default: :linear

opacityresponse: The column which its values will be used to determine the opacity of polygons.

outline: If true an outline will be drawn for each polygon. default: true

Grouping

group: The name of column for grouping observation. Each group of observations will create seperate polygon and polygons in each group will have different color.

Polygon appearance

color: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4"), or pass a gradient color using the gradient() function. default: :steelblue

colormodel: The color model which will be used for fill color when colorresponse is passed. It can be an scheme or a vector of colors. default: :diverging

opacity: The mark opacity from 0 (transparent) to 1 (opaque). default: 1

outlinecolor: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4"), or pass a gradient color using the gradient() function. default: :steelblue

outlinedash: The outline dash style. default: [0]

outlineopacity: The ouline opacity. default: 1

outlinethickness: The outline thickness. default: 1

Axes options

x2axis: When set to true, the top x-axis will be used for the current plot. default: false

y2axis: When set to true, the right y-axis will be used for the current plot. default: false

Legend

legend: User can pass a symbol to this keyword argument to indicate that more customisation will be passed for the legened of corresponding mark. User needs to provide the extra customisation via the Legend global keyword.

Miscellaneous

clip: Indicates if the marks should be clipped.

StatisticalGraphics.RefLine — Type

RefLine(args...)

Represent a Refline with given arguments.

Keyword arguments

Required

axis: WHich axis should be used to draw the reflines, i.e. :xaxis, :x2axis, :yaxis, :y2axis.

values: The position of reflines, it can be a vector of values. default: 0

Refline appearance

color: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4"), or pass a gradient color using the gradient() function. default: :grey

dash: The line dash style. default: [0]

opacity: The mark opacity from 0 (transparent) to 1 (opaque). default: 1

thickness: The line thickness. default: 1

Miscellaneous

clip: Indicates if the marks should be clipped.

StatisticalGraphics.Reg — Type

Reg(args...)

Represent a Regression Line with given arguments.

Keyword arguments

Required

x: The column to be used as x, i.e. the independent variable. default: 0

y: The column to be used as y, i.e. the response variable. default: 0

Reg Options

degree: The maximum polynomial degree for the fitted line. default: 1

intercept: Whether the fitted line should have the intercept term. default: true

interpolate: The interplate function to use for drawing lines, e.g. :linear, :basis, :natural, :step, ... default: :linear

npoints: The number of points in grid for drawing the regression line. default: 100

Grouping

group: The name of column for grouping observation. Each group of observations will create seperate Regression Line.

Reg appearance

color: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4"), or pass a gradient color using the gradient() function. default: "#4682b4"

dash: The Line dash style. default: [0]

opacity: The mark opacity from 0 (transparent) to 1 (opaque). default: 1

thickness: The Line thickness. default: 1

Confidence Interval

alpha: Control the confidence intervals level. default: 0.05

cli: Draw the prediction confidence interval. default: false

clicolor: The filling color for the prediction confidence interval.

cliopacity: The opacity for the prediction confidence interval. default: 0.3

clm: Draw the mean confidence interval. default: false

clmcolor: The filling color for the mean confidence interval.

clmopacity: The opacity for the mean confidence interval. default: 0.3

Axes options

x2axis: When set to true, the top x-axis will be used for the current plot. default: false

y2axis: When set to true, the right y-axis will be used for the current plot. default: false

Legend

legend: User can pass a symbol to this keyword argument to indicate that more customisation will be passed for the legened of corresponding mark. User needs to provide the extra customisation via the Legend global keyword.

Miscellaneous

clip: Indicates if the marks should be clipped.

StatisticalGraphics.Scatter — Type

Scatter(args...)

Represent a Scatter plot with given arguments.

Keyword arguments

Required

x: The column to be used as x coordinate. default: 0

y: The column to be used as y coordinate. default: 0

Scatter Options

angleresponse: The column which its values will be used to determine the angle of the marks.

colorresponse: The column which its values will be used to determine the fill color of the marks.

jitter: The jitter strength in the x and y axes direction, respectively. default: [0, 0]

opacityresponse: The column which its values will be used to determine the opacity of the marks.

symbolresponse: The column which its values will be used to determine the symbol of the marks.

Grouping

group: The name of column for grouping observation. Each group of observations will create seperate scatter plot.

Scatter appearance

angle: The symbol angle. default: 0

color: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4").

colormodel: It specifies the color scheme to use for the marks. default: :diverging

opacity: The mark opacity from 0 (transparent) to 1 (opaque). default: 1

outlinecolor: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4").

size: The symbol size. default: 50

symbol: The symbol type, e.g. :circle, :square, ... default: "circle"

thickness: The mark outline thickness. default: 1

Scatter Label

labelalgorithm: The algorithm for placing labels. default: :naive

labelanchor: The anchors for choosing the location of labels. default: [:top, :bottom, :left, :right]

labelangle: The text angle. default: 0

labelcolor: The label text color, it can also be :group or :colorresponse for choosing the color based on the points groups. default: :black

labeldir: The direction of the text, i.e. :ltr or :rtl. default: :ltr

labelfont: The font name for displaying text.

labelfontweight: The font weight for displaying text, use 100 for thin font and 900 for the bold one.

labelitalic: If true the italic style will be used for displaying text.

labellimit: The maximum length of the text mark in pixels. The text value will be automatically truncated if the rendered size exceeds the limit.

labelresponse: The column which its values will be used to label points.

labelsize: The font size for displaying text.

tooltip: A tooltip will be shown upon mouse hover. default: false

Axes options

x2axis: When set to true, the top x-axis will be used for the current plot. default: false

xshift: Shift the mark in direction of x. Useful for discrete type axes. default: 0

y2axis: When set to true, the right y-axis will be used for the current plot. default: false

yshift: Shift the mark in direction of y. Useful for discrete type axes. default: 0

Legend

legend: User can pass a symbol to this keyword argument to indicate that more customisation will be passed for the legened of corresponding mark. User needs to provide the extra customisation via the Legend global keyword.

Miscellaneous

clip: Indicates if the marks should be clipped.

StatisticalGraphics.Segment — Type

Segment(args...)

Represent a Segment line with given arguments.

Keyword arguments

Required

lower: The lower values for segments. User can pass a column or a Float value. default: 0

upper: The upper values for segments. User can pass a column or a Float value. default: 0

x: The column to be used as x coordinate. User must pass either this or the y argument. default: 0

y: The column to be used as y coordinate. User must pass either this or the x argument. default: 0

Grouping

group: The name of column for grouping observation. Each group of observations will create seperate segment lines.

Segment appearance

color: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4"), or pass a gradient color using the gradient() function. default: "#4682b4"

opacity: The mark opacity from 0 (transparent) to 1 (opaque). default: 1

thickness: The thickness of the mark default: 1

Axes options

x2axis: When set to true, the top x-axis will be used for the current plot. default: false

y2axis: When set to true, the right y-axis will be used for the current plot. default: false

Legend

legend: User can pass a symbol to this keyword argument to indicate that more customisation will be passed for the legened of corresponding mark. User needs to provide the extra customisation via the Legend global keyword.

Miscellaneous

clip: Indicates if the marks should be clipped.

StatisticalGraphics.TextPlot — Type

TextPlot(args...)

Represent a Text plot with given arguments.

Keyword arguments

Required

text: The column to be used as text values for each point.

x: The column to be used as x coordinate.

y: The column to be used as y coordinate.

TextPlot Options

angleresponse: The column which its values will be used to determine the angle of text.

colorresponse: The column which its values will be used to determine the color of text.

opacityresponse: The column which its values will be used to determine the opacity of text.

Grouping

group: The name of column for grouping observation. Each group of observations will create seperate text plot.

TextPlot appearance

align: The text alignment. default: :left

angle: The text angle. default: 0

color: The text color. default: :black

colormodel: It specifies the color scheme to use for the marks. default: :diverging

dir: The direction of the text, i.e. :ltr or :rtl. default: :ltr

font: The font name for displaying text.

fontweight: The font weight for displaying text, use 100 for thin font and 900 for the bold one.

italic: If true the italic style will be used for displaying text.

limit: The maximum length of the text mark in pixels. The text value will be automatically truncated if the rendered size exceeds the limit.

opacity: The mark opacity from 0 (transparent) to 1 (opaque). default: 1

size: The font size for displaying text. default: 10

textbaseline: The text baseline. default: :alphabetic

Axes options

x2axis: When set to true, the top x-axis will be used for the current plot. default: false

y2axis: When set to true, the right y-axis will be used for the current plot. default: false

Legend

legend: User can pass a symbol to this keyword argument to indicate that more customisation will be passed for the legened of corresponding mark. User needs to provide the extra customisation via the Legend global keyword.

Miscellaneous

clip: Indicates if the marks should be clipped.

StatisticalGraphics.Violin — Type

Violin(args...)

Represent a Violin Plot with given arguments.

Keyword arguments

Required

x: User should pass multiple columns for plotting the comparative violin plot, i.e. side by side violin plot for passed columns. User must pass either this or the y argument. default: 0

y: User should pass multiple columns for plotting the comparative violin plot, i.e. side by side violin plot for passed columns. User must pass either this or the x argument. default: 0

Violin options

bw: The kernel bandwidth.

categoryorder: How the category should be ordered, i.e. :ascending, :descending, :data. default: :ascending

interpolate: The line interpolation algorithm for drawing the density line. default: :linear

missingmode: Indicate how to handle missing values in category or group. 0 = nothing, 1 = no missing in category. default: 0

npoints: The number of points for drawing the density line. default: 100

scale: The scale to be used for density estimation, see Density. default: StatisticalGraphics.violin_default_scale

weights: The kernel weight function. default: :gaussian

Grouping

category: A category column which indicates the violin plot must be drawn within each category.

Violin appearance

color: The default color for the mark. User can pass color's name as symbol (e.g. :red), as string (e.g. "red"), as HTML color value (e.g. "#4682b4"), or pass a gradient color using the gradient() function.

filled: Specifiy the plot should be filled with color. default: true

fillopacity: Specifiy The fill opacity of the plot. default: 0.5

groupspace: The space between violins inside each category. It must be a number between 0 and 1. default: 0.05

opacity: Specifiy The opacity of the density line. default: 1

side: Specifiy which half of the violin plot should be plotted, i.e. :right(:bottom), :left(:top). default: :both

space: The space between violins. It must be a number between 0 and 1. default: 0.1

thickness: Specifiy The thickness of the outline line. default: 1

Axes options

x2axis: When set to true, the top x-axis will be used for the current plot. default: false

y2axis: When set to true, the right y-axis will be used for the current plot. default: false

Legend

legend: User can pass a symbol to this keyword argument to indicate that more customisation will be passed for the legened of corresponding mark. User needs to provide the extra customisation via the Legend global keyword.

Miscellaneous

clip: Indicates if the marks should be clipped.